refactor(exports): default to the fetch build, special-case node
Invert the package entry points to mirror get-it's convention: the
platform-neutral fetch build is now the default (`dist/index.js`) and
the undici/Node build is the special case (`dist/index.node.js`).
Previously the fetch build was confusingly named `index.browser.js` even
though deno, bun, workerd, and edge all used it - a `.browser.js` frame
on a non-browser runtime was misleading.

- src: `index.browser.ts` -> `index.ts` (fetch/default, browserMiddleware),
  old `index.ts` -> `index.node.ts` (nodeMiddleware/undici).
- exports `.`: only `node` resolves the Node build; `react-server`,
  `sanity-function`, `deno`, `bun`, `workerd`, `worker` are listed ahead
  of `node` so the runtimes that co-activate the `node` condition still
  get the fetch build. Everything else falls through to `default`.
- exports `./stega`: drop the redundant `browser` condition - it built a
  byte-identical file to `dist/stega.js` (same source, no env branching).
- drop the now-unnecessary top-level `browser` field.
- point the worker/browser/edge test configs at the default source and
  the Node suite at the `node` source.

Work around a rolldown code-splitting bug surfaced by the new entry
graph: the `stegaEncodeSourceMap` chunk (created by a dynamic import in
dataMethods) omitted `encodeIntoResult` from its exports even though
`@sanity/client/stega` re-exports it, breaking that import at runtime.
Re-exporting `encodeIntoResult` from the `stegaEncodeSourceMap` module
forces it into the chunk's export list.
