+++ package.json
@@ -27,21 +27,16 @@
   "exports": {
     ".": {
       "source": "./src/index.ts",
-      "browser": {
-        "source": "./src/index.browser.ts",
-        "import": "./dist/index.browser.js"
+      "react-server": "./dist/index.js",
+      "sanity-function": "./dist/index.js",
+      "deno": "./dist/index.js",
+      "bun": "./dist/index.js",
+      "workerd": "./dist/index.js",
+      "worker": "./dist/index.js",
+      "node": {
+        "source": "./src/index.node.ts",
+        "import": "./dist/index.node.js"
       },
-      "react-native": {
-        "import": "./dist/index.browser.js"
-      },
-      "sanity-function": "./dist/index.browser.js",
-      "react-server": "./dist/index.browser.js",
-      "bun": "./dist/index.browser.js",
-      "deno": "./dist/index.browser.js",
-      "edge": "./dist/index.browser.js",
-      "edge-light": "./dist/index.browser.js",
-      "workerd": "./dist/index.browser.js",
-      "worker": "./dist/index.browser.js",
       "import": "./dist/index.js",
       "default": "./dist/index.js"
     },
@@ -52,10 +47,6 @@
     },
     "./stega": {
       "source": "./src/stega/index.ts",
-      "browser": {
-        "source": "./src/stega/index.ts",
-        "import": "./dist/stega.browser.js"
-      },
       "import": "./dist/stega.js",
       "default": "./dist/stega.js"
     },
@@ -69,9 +60,6 @@
   "main": "./dist/index.js",
   "unpkg": "umd/sanityClient.min.js",
   "module": "./dist/index.js",
-  "browser": {
-    "./dist/index.js": "./dist/index.browser.js"
-  },
   "types": "./dist/index.d.ts",
   "typesVersions": {
     "*": {
