+++ runtimes/deno/client_fetch_test.mjs
@@ -1,5 +1,5 @@
 import {assert} from 'https://deno.land/std@0.171.0/testing/asserts.ts'
-import {createClient} from '../../dist/index.browser.js'
+import {createClient} from '../../dist/index.js'
 
 const projectId = '81pocpw8'
 const dataset = 'production'
