+++ src/index.node.ts
@@ -1,6 +1,6 @@
 import defineCreateClientExports, {type ClientConfig, SanityClient} from './defineCreateClient'
 import {defineDeprecatedCreateClient} from './defineDeprecatedCreateClient'
-import environment from './http/browserMiddleware'
+import environment from './http/nodeMiddleware'
 import type {Requester} from './types'
 
 export * from './defineCreateClient'
@@ -10,7 +10,11 @@ const exp = defineCreateClientExports<SanityClient, ClientConfig>(environment, S
 /** @public */
 export const requester: Requester = exp.requester
 
-/** @public */
+/**
+ * @remarks
+ * As of API version `v2025-02-19`, the default perspective used by the client has changed from `raw` to `published`. {@link https://www.sanity.io/changelog/676aaa9d-2da6-44fb-abe5-580f28047c10|Changelog}
+ * @public
+ */
 export const createClient = exp.createClient
 
 /**
