+++ vitest.browser.config.ts
@@ -13,9 +13,8 @@ export default mergeConfig(
       environment: 'happy-dom',
       alias: {
         '@sanity/client/csm': new URL(pkg.exports['./csm'].source, import.meta.url).pathname,
-        '@sanity/client/stega': new URL(pkg.exports['./stega'].browser.source, import.meta.url)
-          .pathname,
-        '@sanity/client': new URL(pkg.exports['.'].browser.source, import.meta.url).pathname,
+        '@sanity/client/stega': new URL(pkg.exports['./stega'].source, import.meta.url).pathname,
+        '@sanity/client': new URL(pkg.exports['.'].source, import.meta.url).pathname,
       },
       typecheck: {
         enabled: false,
